토큰 -> id 검증 구현

+++ bulletin-board/src/main/java/com/solo/bulletin_board/member/service/MemberService.java
@@ -1,5 +1,6 @@
 package com.solo.bulletin_board.member.service;
 
+import com.solo.bulletin_board.auth.userDetailsService.CustomUserDetails;
 import com.solo.bulletin_board.auth.utils.CustomAuthorityUtils;
 import com.solo.bulletin_board.exception.BusinessLogicException;
 import com.solo.bulletin_board.exception.ExceptionCode;
@@ -42,6 +43,14 @@ public class MemberService {
         }
     }
 
+    public void checkMemberId(long memberId, CustomUserDetails customUserDetails){
+
+        if(memberId != customUserDetails.getMemberId()){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_NOT_MATCHED);
+        }
+
+    }
+
     public Member findVerifiedMember(long memberId){
 
         Optional<Member> optionalMember = memberRepository.findById(memberId);
@@ -52,6 +61,7 @@ public class MemberService {
         return findMember;
 
     }
+
 
     public Member createMember(Member member){
         verifyExistsEmail(member.getEmail());
@@ -66,8 +76,8 @@ public class MemberService {
         return memberRepository.save(member);
     }
 
-    public Member updateMember(Member member){
-        Member findMember = findVerifiedMember(member.getMemberId());
+    public Member updateMember(Member member, CustomUserDetails customUserDetails){
+        Member findMember = findVerifiedMember(customUserDetails.getMemberId());
 
         Optional.ofNullable(member.getNickname())
                 .ifPresent(nickname -> {
@@ -81,13 +91,18 @@ public class MemberService {
 
     }
 
-    public Member findMember(long memberId){
-        return findVerifiedMember(memberId);
+    public Member findMember(CustomUserDetails customUserDetails){
+        return findVerifiedMember(customUserDetails.getMemberId());
     }
 
-    public void deleteMember(long memberId){
+    public void deleteMember(String password, CustomUserDetails customUserDetails){
 
-        Member findMember = findVerifiedMember(memberId);
+        Member findMember = findVerifiedMember(customUserDetails.getMemberId());
+
+        if(!passwordEncoder.matches(password, findMember.getPassword())){
+            throw new BusinessLogicException(ExceptionCode.MEMBER_PASSWORD_NOT_MATCHED);
+        }
+
         memberRepository.delete(findMember);
     }
